backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==4.2.11
django-allauth==64.2.1
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.3
django-bootstrap5==24.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
marshmallow==3.21.1
outcome==1.3.0.post0
packaging==24.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-core==2.18.1
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing-extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
